added deflate and br decompress to http 1.1
