+++ http2/transport.go
@@ -9,9 +9,6 @@ package http2
 import (
 	"bufio"
 	"bytes"
-	"compress/flate"
-	"compress/gzip"
-	"compress/zlib"
 	"context"
 	"crypto/rand"
 	"crypto/tls"
@@ -30,8 +27,6 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
-
-	"github.com/andybalholm/brotli"
 
 	http "github.com/useflyent/fhttp"
 	"github.com/useflyent/fhttp/httptrace"
@@ -2206,29 +2201,8 @@ func (rl *clientConnReadLoop) handleResponse(cs *clientStream, f *MetaHeadersFra
 	res.Body = transportResponseBody{cs}
 	go cs.awaitRequestCancel(cs.req)
 
-	res.Body = DecompressBody(res)
+	res.Body = http.DecompressBody(res)
 	return res, nil
-}
-
-func DecompressBody(res *http.Response) io.ReadCloser {
-	ce := res.Header.Get("Content-Encoding")
-	res.ContentLength = -1
-	res.Uncompressed = true
-
-	switch ce {
-	case "gzip":
-		return &gzipReader{
-			body: res.Body,
-		}
-	case "br":
-		return &brReader{
-			body: res.Body,
-		}
-	case "deflate":
-		return identifyDeflate(res.Body)
-	default:
-		return res.Body
-	}
 }
 
 func (rl *clientConnReadLoop) processTrailers(cs *clientStream, f *MetaHeadersFrame) error {
@@ -2789,140 +2763,6 @@ type erringRoundTripper struct{ err error }
 
 func (rt erringRoundTripper) RoundTripErr() error                             { return rt.err }
 func (rt erringRoundTripper) RoundTrip(*http.Request) (*http.Response, error) { return nil, rt.err }
-
-// gzipReader wraps a response body so it can lazily
-// call gzip.NewReader on the first call to Read
-type gzipReader struct {
-	_    incomparable
-	body io.ReadCloser // underlying Response.Body
-	zr   *gzip.Reader  // lazily-initialized gzip reader
-	zerr error         // sticky error
-}
-
-func (gz *gzipReader) Read(p []byte) (n int, err error) {
-	if gz.zerr != nil {
-		return 0, gz.zerr
-	}
-	if gz.zr == nil {
-		gz.zr, err = gzip.NewReader(gz.body)
-		if err != nil {
-			gz.zerr = err
-			return 0, err
-		}
-	}
-	return gz.zr.Read(p)
-}
-
-func (gz *gzipReader) Close() error {
-	return gz.body.Close()
-}
-
-// brReader lazily wraps a response body into an
-// io.ReadCloser, will call gzip.NewReader on first
-// call to read
-type brReader struct {
-	_    incomparable
-	body io.ReadCloser
-	zr   *brotli.Reader
-	zerr error
-}
-
-func (br *brReader) Read(p []byte) (n int, err error) {
-	if br.zerr != nil {
-		return 0, br.zerr
-	}
-	if br.zr == nil {
-		br.zr = brotli.NewReader(br.body)
-	}
-	return br.zr.Read(p)
-}
-
-func (br *brReader) Close() error {
-	return br.body.Close()
-}
-
-type zlibDeflateReader struct {
-	_    incomparable
-	body io.ReadCloser
-	zr   io.ReadCloser
-	err  error
-}
-
-func (z *zlibDeflateReader) Read(p []byte) (n int, err error) {
-	if z.err != nil {
-		return 0, z.err
-	}
-	if z.zr == nil {
-		z.zr, err = zlib.NewReader(z.body)
-		if err != nil {
-			z.err = err
-			return 0, z.err
-		}
-	}
-	return z.zr.Read(p)
-}
-
-func (z *zlibDeflateReader) Close() error {
-	return z.zr.Close()
-}
-
-type deflateReader struct {
-	_    incomparable
-	body io.ReadCloser
-	r    io.ReadCloser
-	err  error
-}
-
-func (dr *deflateReader) Read(p []byte) (n int, err error) {
-	if dr.err != nil {
-		return 0, dr.err
-	}
-	if dr.r == nil {
-		dr.r = flate.NewReader(dr.body)
-	}
-	return dr.r.Read(p)
-}
-
-func (dr *deflateReader) Close() error {
-	return dr.r.Close()
-}
-
-const (
-	zlibMethodDeflate = 0x78
-	zlibLevelDefault  = 0x9C
-	zlibLevelLow      = 0x01
-	zlibLevelMedium   = 0x5E
-	zlibLevelBest     = 0xDA
-)
-
-func identifyDeflate(body io.ReadCloser) io.ReadCloser {
-	var header [2]byte
-	_, err := io.ReadFull(body, header[:])
-	if err != nil {
-		return body
-	}
-
-	if header[0] == zlibMethodDeflate &&
-		(header[1] == zlibLevelDefault || header[1] == zlibLevelLow || header[1] == zlibLevelMedium || header[1] == zlibLevelBest) {
-		return &zlibDeflateReader{
-			body: prependBytesToReadCloser(header[:], body),
-		}
-	} else if header[0] == zlibMethodDeflate {
-		return &deflateReader{
-			body: prependBytesToReadCloser(header[:], body),
-		}
-	}
-	return body
-}
-
-func prependBytesToReadCloser(b []byte, r io.ReadCloser) io.ReadCloser {
-	w := new(bytes.Buffer)
-	w.Write(b)
-	io.Copy(w, r)
-	defer r.Close()
-
-	return io.NopCloser(w)
-}
 
 type errorReader struct{ err error }
